requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
dj-rest-auth==5.0.2
Django==4.2.7
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
Pillow==10.1.0
psycopg2-binary==2.9.9
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
urllib3==2.1.0
whitenoise==6.6.0
